Fix DiedBottomSheet
